updated pgcps_functions.js, index.html and package.json
In pgcps_functions, I created a function to run other functions when the as soon as the webpage loads. In pgcps_functions I changed the api endpoints to the appropriate AWS API Gateway endpoints and updated some of the code to ES6 standards. In index moved the onload property written into the body tag to the pgcps_functions. In package.json I added "type: "module".